fix(onboarding): move tracking of app start
SVA-305
