+++ src/helpers/initializationHelper.ts
@@ -18,8 +18,6 @@ const initializeLocationServices = async () => {
   }
 };
 
-// TODO: check how to do it with matomo and initialization differences (pre/post rendering the provider)
-// -> change type from `() => void` to `(onboarding?: boolean) => void` ?
 export const Initializers = {
   [Initializer.LocationService]: initializeLocationServices,
   [Initializer.MatomoTracking]: showMatomoAlert,
